Separated views and controllers, separated threads
