+++ pb_tool.cfg
@@ -64,7 +64,7 @@ extras: metadata.txt
 
 # Other directories to be deployed with the plugin.
 # These must be subdirectories under the plugin directory
-extra_dirs: ui assets models utils
+extra_dirs: models views controllers threads assets utils
 
 # ISO code(s) for any locales (translations), separated by spaces.
 # Corresponding .ts files must exist in the i18n directory
