+++ views/collection_loading_dialog.ui
