+++ views/query_dialog.ui
