+++ views/results_dialog.ui
