+++ views/select_bands_dialog.ui
